# apps/ideploy/openapi.yaml
openapi: 3.1.0
info:
  title: Coolify
  version: "0.1"
servers:
  - url: "https://app.ideploy.io/api/v1"
    description: "Coolify Cloud API. Change the host to your own instance if you are self-hosting."
paths:
...
  /enable:
    get:
      summary: "Enable API"
      description: "Enable API (only with root permissions)."
      operationId: enable-api
      responses:
        "200":
          description: "Enable API."
          content:
            application/json:
              schema:
                properties:
                  message: { type: string, example: "API enabled." }
                type: object
        "403":
          description: "You are not allowed to enable the API."
          content:
            application/json:
              schema:
                properties:
                  message:
                    {
                      type: string,
                      example: "You are not allowed to enable the API.",
                    }
                type: object
        "401":
          $ref: "#/components/responses/401"
        "400":
          $ref: "#/components/responses/400"
      security:
        - bearerAuth: []
  /disable:
    get:
      summary: "Disable API"
      description: "Disable API (only with root permissions)."
      operationId: disable-api
      responses:
        "200":
          description: "Disable API."
          content:
            application/json:
              schema:
                properties:
                  message: { type: string, example: "API disabled." }
                type: object
        "403":
          description: "You are not allowed to disable the API."
          content:
            application/json:
              schema:
                properties:
                  message:
                    {
                      type: string,
                      example: "You are not allowed to disable the API.",
                    }
                type: object
        "401":
          $ref: "#/components/responses/401"
        "400":
          $ref: "#/components/responses/400"
      security:
        - bearerAuth: []
...
components:
...
  responses:
    "400":
      description: "Invalid token."
      content:
        application/json:
          schema:
            properties:
              message:
                type: string
                example: "Invalid token."
            type: object
    "401":
      description: Unauthenticated.
      content:
        application/json:
          schema:
            properties:
              message:
                type: string
                example: Unauthenticated.
            type: object
...
  securitySchemes:
    bearerAuth:
      type: http
      description: "Go to `Keys & Tokens` / `API tokens` and create a new token. Use the token as the bearer token."
      scheme: bearer
...